+++ NEWS.md
@@ -2,6 +2,7 @@
 # variancePartition 1.23.2
 	- add flag to checkModelStatus() so warnings are thrown immediately
 	- fix export of as.data.frame
+	- fixed issues where messages were printed even if quiet=TRUE
 
 # variancePartition 1.23.1
 	- New freeze for Bioconductor devel branch